Move withViewport and Viewport to separate files
